backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
vaderSentiment==3.3.2
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
